Content-generator_tool/Backend/requirements.txt
flask
requests
python-dotenv
flask-cors

reportlab
google-generativeai